requirements.txt
# requirements.txt
numpy>=1.21.0
librosa>=0.9.0
openai-whisper>=20230314
scipy>=1.7.0
matplotlib>=3.5.0
ffmpeg-python>=0.2.0
